Refactor version manager and Add `versioner`
